小功能更新：1、float型改为实际用double型存储（float类型仅仅标识是小数类型） 2、忽略对Excel自动生成的以~$开头的临时文件的处理

+++ XlsxToLua/TableAnalyzeHelper.cs
@@ -527,8 +527,8 @@ public class TableAnalyzeHelper
                 }
                 else
                 {
-                    float floatValue;
-                    bool isValid = float.TryParse(inputData, out floatValue);
+                    double floatValue;
+                    bool isValid = double.TryParse(inputData, out floatValue);
                     if (isValid)
                         fieldInfo.Data.Add(floatValue);
                     else
